pyproject.toml
[tool.poetry]
name = "paul-chess-pos-lichess"
version = "0.1.0"
description = ""
authors = ["ChurroC <charanvenkatchandran@gmail.com>"]
readme = "README.md"
packages = [{include = "paul_chess_pos_lichess"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
img2table = "^1.2.5"
opencv-python = "^4.8.1.78"
pillow = "^10.1.0"

[tool.poetry.scripts]
#poetry run python paul_chess_pos_lichess/main.py

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
